+++ data/src/main/java/com/shipsmart/data/UserRepository.kt
@@ -1,0 +1,4 @@
+package com.shipsmart.data
+
+class UserRepository {
+}
