+++ data/src/main/java/com/shipsmart/data/dbLogic/AndroidDB.kt
@@ -1,4 +1,4 @@
-package com.shipsmart.data.dbLogic
+package shipsmart.dbLogic
 
 import android.content.Context
 import androidx.room.Database
