+++ domain/src/main/java/com/shipsmart/domain/model/RegistrationParams.kt
@@ -1,3 +1,3 @@
-package com.shipsmart.domain.model
+package com.example.domain.model
 
 class RegistrationParams(val email: String, val password: String)
