+++ domain/src/main/java/com/shipsmart/domain/model/SignActivityObjects.kt
@@ -1,0 +1,6 @@
+package com.shipsmart.domain.model
+
+import android.widget.Button
+import android.widget.TextView
+
+class SignActivityObjects(val activityLabel: TextView, val enterButton: Button, val changeActivityButton: TextView)
