+++ domain/src/main/java/com/shipsmart/domain/model/SupabaseUser.kt
