+++ domain/src/main/java/com/shipsmart/domain/repository/DBdao.kt
@@ -1,6 +1,6 @@
-package com.shipsmart.domain.repository
+package com.example.domain.repository
 
-import com.shipsmart.domain.model.SupabaseUser
+import com.example.domain.model.SupabaseUser
 
 interface DBdao {
     suspend fun getUser(email: String) : SupabaseUser?
