+++ domain/src/main/java/com/shipsmart/domain/repository/ToastConstructor.kt
@@ -1,4 +1,4 @@
-package com.shipsmart.domain.repository
+package com.example.domain.repository
 
 import android.content.Context
 import android.util.Log
