+++ domain/src/main/java/com/shipsmart/domain/repository/UserRepositoryInterface.kt
@@ -1,0 +1,4 @@
+package com.shipsmart.domain.repository
+
+class UserRepositoryInterface {
+}
